feat: VisualMediaGalleryArranger
