+++ src/dotnet/UI.Blazor.App/Components/ChatView/Items/Attachment/VisualMediaGalleryArranger/VisualMediaProportions.cs
@@ -1,0 +1,8 @@
+namespace ActualChat.UI.Blazor.App.Components;
+
+public enum VisualMediaProportions {
+    Narrow,
+    Square,
+    Wide,
+    ExtraWide,
+}
