+++ src/dotnet/UI.Blazor.App/Components/ChatView/Items/Attachment/VisualUploadMediaLine.razor
@@ -17,7 +17,7 @@
 @code {
     [CascadingParameter] public ScreenSize ScreenSize { get; set; }
 
-    [Parameter, EditorRequired] public MediaGalleryLine<Attachment> Row { get; set; } = null!;
+    [Parameter, EditorRequired] public VisualMediaGalleryLine<Attachment> Row { get; set; } = null!;
     [Parameter, EditorRequired] public IAttachmentList Uploads { get; set; } = null!;
     [Parameter] public Attachment[]? MediaList { get; set; }
     [Parameter] public string Class { get; set; } = "";
